rearrage: move HTML related utils into HTMLUtils
Move HTML related utility functions into its own file to allow
non-HTML related functions to be saved inside folders with 'ts'
extension instead of 'tsx'.
